backend/requirements.txt
django
psycopg2-binary
djangorestframework
django-storages
boto3
django-cors-headers
